패키징 재처리 툴팁추가

+++ TangoQ/app/src/main/java/com/tangoplus/tangoq/data/MessageVO.kt
@@ -2,7 +2,7 @@ package com.tangoplus.tangoq.data
 
 
 data class MessageVO(
-    val sn: Long = 0L,
+    val sn: Int = 0 ,
     val message : String = "",
     val timeStamp : Long = 0L,
     val route : String = ""
